+++ sql/tax_adjustments.sql
@@ -3,7 +3,7 @@ create table tax_adjustments(
   worker varchar(40),
   completion_year varchar(4) not null,
   payment_finalization_year varchar(4) not null,
-  adjustment float(7,2) not null,
+  adjustment decimal(7,2) not null,
   constraint adjustment_is_for_worker foreign key (worker) references workers (worker)
  ) ENGINE=InnoDB AUTO_INCREMENT=1 DEFAULT CHARSET=utf8 COLLATE=utf8_unicode_ci;
